+++ iqdb_tagger/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 """Init file."""
-__version__ = '0.2.2'
+__version__ = '0.3.0'
 db_version = 1