+++ tests/test_main.py
@@ -7,8 +7,9 @@ import vcr
 from click.testing import CliRunner
 from PIL import Image
 
-from iqdb_tagger.__main__ import get_posted_image, init_program, run
-from iqdb_tagger.models import ImageModel, ThumbnailRelationship
+from iqdb_tagger.__main__ import cli_run, init_program
+from iqdb_tagger.models import ImageModel, ThumbnailRelationship, get_posted_image
+import iqdb_tagger
 
 logging.basicConfig()
 vcr_log = logging.getLogger('vcr')
@@ -69,8 +70,9 @@ def test_main(tmpdir, tmp_img):  # pylint:disable=redefined-outer-name
     db_path = tmpdir.join('temp_db.db')
 
     runner = CliRunner()
+    os.environ['FLASK_APP'] = iqdb_tagger.__main__.__file__
     result = runner.invoke(
-        run, [
+        cli_run, [
             '--db-path', db_path.strpath, '--resize',
             '--match-filter', 'best-match', img_path.strpath
         ]